requirements/common.txt
Django==1.9.11
djangorestframework==3.6.2
wheel==0.24.0
pytz==2017.2
django-cors-headers==2.0.2
